+++ crates/miden-testing/src/kernel_tests/tx/test_epilogue.rs
@@ -54,12 +54,12 @@ fn test_epilogue() -> anyhow::Result<()> {
     let account = Account::mock(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE, Auth::IncrNonce);
     let tx_context = {
         let output_note_1 =
-            create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), &[FungibleAsset::mock(100)]);
+            create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), [FungibleAsset::mock(100)]);
 
         // input_note_1 is needed for maintaining cohesion of involved assets
         let input_note_1 =
-            create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), &[FungibleAsset::mock(100)]);
-        let input_note_2 = create_spawn_note(ACCOUNT_ID_SENDER.try_into()?, vec![&output_note_1])?;
+            create_p2any_note(ACCOUNT_ID_SENDER.try_into().unwrap(), [FungibleAsset::mock(100)]);
+        let input_note_2 = create_spawn_note([&output_note_1])?;
         TransactionContextBuilder::new(account.clone())
             .extend_input_notes(vec![input_note_1, input_note_2])
             .extend_expected_output_notes(vec![OutputNote::Full(output_note_1)])
@@ -153,12 +153,12 @@ fn test_compute_output_note_id() -> anyhow::Result<()> {
         let account =
             Account::mock(ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_UPDATABLE_CODE, Auth::IncrNonce);
         let output_note_1 =
-            create_p2any_note(ACCOUNT_ID_SENDER.try_into()?, &[FungibleAsset::mock(100)]);
+            create_p2any_note(ACCOUNT_ID_SENDER.try_into()?, [FungibleAsset::mock(100)]);
 
         // input_note_1 is needed for maintaining cohesion of involved assets
         let input_note_1 =
-            create_p2any_note(ACCOUNT_ID_SENDER.try_into()?, &[FungibleAsset::mock(100)]);
-        let input_note_2 = create_spawn_note(ACCOUNT_ID_SENDER.try_into()?, vec![&output_note_1])?;
+            create_p2any_note(ACCOUNT_ID_SENDER.try_into()?, [FungibleAsset::mock(100)]);
+        let input_note_2 = create_spawn_note([&output_note_1])?;
         TransactionContextBuilder::new(account)
             .extend_input_notes(vec![input_note_1, input_note_2])
             .extend_expected_output_notes(vec![OutputNote::Full(output_note_1)])
@@ -235,7 +235,7 @@ fn test_epilogue_asset_preservation_violation_too_few_input() -> anyhow::Result<
         .dynamically_linked_libraries(TransactionKernel::mock_libraries())
         .build()?;
 
-    let input_note = create_spawn_note(account.id(), vec![&output_note_1, &output_note_2])?;
+    let input_note = create_spawn_note([&output_note_1, &output_note_2])?;
 
     let tx_context = mock_chain
         .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?
@@ -309,10 +309,7 @@ fn test_epilogue_asset_preservation_violation_too_many_fungible_input() -> anyho
         .dynamically_linked_libraries(TransactionKernel::mock_libraries())
         .build()?;
 
-    let input_note = create_spawn_note(
-        ACCOUNT_ID_SENDER.try_into()?,
-        vec![&output_note_1, &output_note_2, &output_note_3],
-    )?;
+    let input_note = create_spawn_note([&output_note_1, &output_note_2, &output_note_3])?;
 
     let tx_context = mock_chain
         .build_tx_context(TxContextInput::AccountId(account.id()), &[], &[input_note])?